Fixed: incorrect statements also got appended to code variable [only works for token errors]
Before, no matter if the code was correct or not it got appended into the code variable. This is a problem cause the wrong code also passes through gets executed in the py_parser. I've added a boolean variable error which tell if there's an error in the code. This is updated at every line. error boolean is used to check whether the current line should be added to the code variable or not.

IT ONLY WORKS FOR  TOKEN ERRORS FOR NOW. WE NEED TO UPDATE IT FOR PRODUCTION/SYNTAX ERRORS AS WELL.